Insert limited fields if limited fields set
