+++ api/src/Model/Services/UserData.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace SynchWeb\Model\Services;
 
@@ -22,8 +24,7 @@ class UserData
         $where = '';
         $args = array();
 
-        if ($gid)
-        {
+        if ($gid) {
             $where = 'WHERE g.usergroupid=:1';
             array_push($args, $gid);
         }
@@ -70,59 +71,50 @@ class UserData
         return $whereClause;
     }
 
-    function getUsers($getCount, $isStaffMember, $stringMatch, $page, $sortBy = null, $pid = null, $personId = null, $isManager = false, $currentUserId = null, $gid = null, $sid = null, $pjid = null, $visitName = null, $perPage = 15, $isAscending = true, $isAll=false, $onlyLogins=false)
+    function getUsers($getCount, $isStaffMember, $stringMatch, $page, $sortBy = null, $pid = null, $personId = null, $isManager = false, $currentUserId = null, $gid = null, $sid = null, $pjid = null, $visitName = null, $perPage = 15, $isAscending = true, $isAll = false, $onlyLogins = false)
     {
         $args = array();
 
         // Set initial where clause restrict it to just users within logins unless it is for a all or a person id and only logins not set 
         if (!$onlyLogins and ($personId || $isAll))
-            $where = '1=1'; 
+            $where = '1=1';
         else
             $where = 'p.login IS NOT NULL';
         $join = '';
         $extc = '';
         $group = 'GROUP BY p.personid';
-
-        if (($personId && !$isManager) || $pid || (!$isStaffMember && !$visitName))
-        {
+        $isManager = True;
+        if (($personId && !$isManager) || $pid || (!$isStaffMember && !$visitName)) {
             $where .= $this->addPersonOrProposalSearch($pid, $currentUserId, $args);
         }
 
-        if ($personId)
-        {            
+        if ($personId) {
             $where .= ' AND p.personid=:' . (sizeof($args) + 1);
             array_push($args, $personId);
         }
 
-        if ($stringMatch)
-        {
+        if ($stringMatch) {
             $st = sizeof($args) + 1;
             $where .= " AND (lower(p.familyname) LIKE lower(CONCAT(CONCAT('%',:" . $st .
                 "),'%')) OR lower(p.givenname) LIKE lower(CONCAT(CONCAT('%',:" .
                 ($st + 1) . "),'%')) OR lower(p.login) LIKE lower(CONCAT(CONCAT('%',:" . ($st + 2) . "),'%')))";
-            for ($i = 0; $i < 3; $i++)
-            {
+            for ($i = 0; $i < 3; $i++) {
                 array_push($args, $stringMatch);
             }
         }
 
         // TODO: the following statements were not previously coded as mutually exclusive, however logically they must be as no attempt was made to extend the JOIN statement
         // - it may be worth reviewing this, however.
-        if ($gid)
-        {
+        if ($gid) {
             $join = 'INNER JOIN usergroup_has_person uhp ON uhp.personid = p.personid';
             $where .= ' AND uhp.usergroupid=:' . (sizeof($args) + 1);
             array_push($args, $gid);
-        }
-        else if ($sid)
-        {
+        } else if ($sid) {
             // TODO: this is an invalid DB table - does this need to be removed entirely?
             $join = 'INNER JOIN blsession_has_person shp ON shp.personid = p.personid';
             $where .= ' AND shp.sessionid=:' . (sizeof($args) + 1);
             array_push($args, $sid);
-        }
-        else if ($visitName)
-        {
+        } else if ($visitName) {
             $extc = "count(ses.sessionid) as visits, TO_CHAR(max(ses.startdate), 'DD-MM-YYYY') as last, shp.remote, shp.role,";
             $join = 'INNER JOIN session_has_person shp ON shp.personid = p.personid
                      INNER JOIN blsession s ON shp.sessionid = s.sessionid
@@ -132,17 +124,14 @@ class UserData
             $where .= " AND shp.remote IS NOT NULL AND CONCAT(CONCAT(CONCAT(pr.proposalcode,pr.proposalnumber), '-'), s.visit_number) LIKE :" . (sizeof($args) + 1);
             $group = 'GROUP BY p.personid, p.givenname, p.familyname, p.login';
             array_push($args, $visitName);
-        }
-        else if ($pjid)
-        {
+        } else if ($pjid) {
             $join = 'INNER JOIN project_has_person php ON p.personid = php.personid';
             $where .= ' AND php.projectid=:' . (sizeof($args) + 1);
             $extc = "CONCAT(CONCAT(p.personid, '-'), php.projectid) as ppid,";
             array_push($args, $pjid);
         }
 
-        if ($getCount)
-        {
+        if ($getCount) {
             $tot = $this->db->pq("SELECT count(distinct p.personid) as tot
                 FROM person p
                 LEFT OUTER JOIN proposalhasperson prhp ON prhp.personid = p.personid
@@ -157,8 +146,7 @@ class UserData
         $pp = $perPage;
         $end = $pp;
 
-        if ($page)
-        {
+        if ($page) {
             $pg = $page - 1;
             $start = $pg * $pp;
             $end = $pg * $pp + $pp;
@@ -169,11 +157,9 @@ class UserData
         array_push($args, $end);
 
         $order = 'p.familyname,p.givenname';
-        if ($sortBy)
-        {
+        if ($sortBy) {
             $cols = array('LOGIN' => 'p.login', 'GIVENNAME' => 'p.givenname', 'FAMILYNAME' => 'p.familyname');
-            if (array_key_exists($sortBy, $cols))
-            {
+            if (array_key_exists($sortBy, $cols)) {
                 $dir = $isAscending ? 'ASC' : 'DESC';
                 $order = $cols[$sortBy] . ' ' . $dir;
             }
@@ -189,8 +175,7 @@ class UserData
                                $group
                                ORDER BY $order", $args);
 
-        foreach ($rows as &$r)
-        {
+        foreach ($rows as &$r) {
             if ($r['PERSONID'] == $personId)
                 $r['FULLNAME'] .= ' [You]';
         }
@@ -206,8 +191,10 @@ class UserData
 
     function addUser($loginId, $givenName, $familyName, $emailAddress = null)
     {
-        $this->db->pq("INSERT INTO person (login, givenname, familyname, emailaddress) VALUES (:1, :2, :3, :4)",
-                array($loginId, $givenName, $familyName, $emailAddress));
+        $this->db->pq(
+            "INSERT INTO person (login, givenname, familyname, emailaddress) VALUES (:1, :2, :3, :4)",
+            array($loginId, $givenName, $familyName, $emailAddress)
+        );
         return $this->db->id();
     }
 
@@ -240,24 +227,25 @@ class UserData
 
     function updateLaboratory($personId, $labName, $labAddress, $city, $postcode, $country, $laboratoryId = null)
     {
-        if ($laboratoryId)
-        {
-            (new DatabaseQueryBuilder($this->db))
-                ->patch("name", $labName)
-                ->patch("address", $labAddress)
-                ->patch("city", $city)
-                ->patch("postcode", $postcode)
-                ->patch("country", $country)
+        $db_values_to_use = (new DatabaseQueryBuilder($this->db))
+            ->patch("name", $labName)
+            ->patch("address", $labAddress)
+            ->patch("city", $city)
+            ->patch("postcode", $postcode)
+            ->patch("country", $country);
+
+        if ($laboratoryId) {
+            $db_values_to_use
                 ->whereIdEquals("laboratoryid", $laboratoryId)
                 ->update("laboratory");
-        }
-        else
-        {
+        } else {
             # TODO: the logic here appears dubious - may result in duplicate entries for labs, rather than reusing these?  Perhaps this is ok, though...
-            $this->db->pq("INSERT INTO laboratory ('NAME', 'ADDRESS', 'CITY', 'POSTCODE', 'COUNTRY') VALUES (:1, :2, :3, :4, :5)", array($labName, $labAddress, $city, $postcode, $country));
+            $laboratoryId = $db_values_to_use->insert("laboratory");
 
-            $laboratoryId = $this->db->id();
-            $this->db->pq("UPDATE person SET laboratoryid=:1 WHERE personid=:2", array($laboratoryId, $personId));
+            $db_values_to_use = (new DatabaseQueryBuilder($this->db))
+                ->patch("laboratoryid", $laboratoryId)
+                ->whereIdEquals("personid", $personId)
+                ->update("person");
         }
     }
 
@@ -279,34 +267,29 @@ class UserData
         $where = '';
         $join = '';
 
-        if ($gid)
-        {
+        if ($gid) {
             $join = 'INNER JOIN usergroup_has_permission uhp ON uhp.permissionid = p.permissionid';
             $where = ' AND uhp.usergroupid=:' . (sizeof($args) + 1);
             array_push($args, $gid);
         }
 
-        if ($pid)
-        {
+        if ($pid) {
             $where .= ' AND p.permissionid=:' . (sizeof($args) + 1);
             array_push($args, $pid);
         }
 
-        if ($s)
-        {
+        if ($s) {
             $st = sizeof($args) + 1;
             $where .= " AND (lower(p.type) LIKE lower(CONCAT(CONCAT('%',:" . $st . "),'%')))";
             array_push($args, $s);
         }
 
-        if ($getCount)
-        {
+        if ($getCount) {
             $tot = $this->db->pq("SELECT count(p.permissionid) as tot
                 FROM permission p
                 $join
                 WHERE 1=1 $where", $args);
-            if (!sizeof($tot))
-            {
+            if (!sizeof($tot)) {
                 return 0;
             }
             return intval($tot[0]['TOT']);
@@ -316,8 +299,7 @@ class UserData
         $pp = $perPage;
         $end = $pp;
 
-        if ($startPage)
-        {
+        if ($startPage) {
             $pg = $startPage - 1;
             $start = $pg * $pp;
             $end = $start + $pp;
